separated pages into a folder
