+++ lib/pages/episodesLandingPage.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/src/foundation/key.dart';
+import 'package:flutter/src/widgets/framework.dart';
+
+class EpisodesLandingPage extends StatelessWidget {
+  const EpisodesLandingPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Text('hehe');
+  }
+}
